Make load balancer ip and annotations configurable
Closes: https://github.com/Quansight/qhub/issues/638

Tried to do the minimal ammount to support load balancer
configuration. One downside I see with this PR is we are introducing
another 2 configuration variables to the qhub_config.yaml. On the
other hand this has been extremely easy to implment thanks to having
the qhub modules in the same directory.